# Conflicts:
#	Monopoly/GradleProject/Monopoly/app/src/main/java/com/rob/monopoly/MainActivity.java

+++ Monopoly/GradleProject/Monopoly/app/src/main/java/com/rob/monopoly/Interfaces/Improvable.java
@@ -9,7 +9,7 @@ public interface Improvable extends Ownable {
     boolean buildHouse();
     boolean demolishHouse();
     int getHousePrice();
-    double getNumHouses();
+    int getNumHouses();
     String getColourGroup();
     boolean ownsColourGroup();
 }
